Add parser for team scores on backend
This adds parser for result of `sv_cmd printstats` command. Result of
this command should be adjusted on xonotic end as well (since it's
output format depends if it was writing into eventlog or console).
When output format would be uniform it could be used to get detail
scores during the game (from the web, or via irc/discord).